backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic_settings
python-jose
passlib
python-multipart